best way to understand what's going on is to convert to typescript. Lots of cleanup
