+++ package.json
@@ -6,6 +6,7 @@
   "version": "0.1.0",
   "author": "Atharv Damle",
   "dependencies": {
+    "@types/react-helmet": "^6.1.0",
     "gatsby": "^2.24.47",
     "gatsby-image": "^2.4.16",
     "gatsby-plugin-dark-mode": "^1.1.2",
