+++ src/components/sidebar.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 
 const Sidebar = ({ header }) => {
-  // console.log({ header });
+  console.log({ header });
   return (
     <div
       className="md:h-screen p-4 cover-content"
